rename madel to superposedpulses
