+++ superposedpulses/example.py
@@ -1,9 +1,9 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-import model.forcing as frc
-import model.point_model as pm
-import model.pulse_shape as ps
+import superposedpulses.forcing as frc
+import superposedpulses.point_model as pm
+import superposedpulses.pulse_shape as ps
 
 # Simplest case, using defaults: exponential pulse shape, exponentially distributed amplitudes, constant duration times.
