+++ superposedpulses/forcing.py
